// src/types/mlb-standings.ts
export const DIVISIONS = {
  "AL East": [1, 10, 2, 30, 14],
  "AL Central": [4, 5, 6, 7, 9],
  "AL West": [11, 18, 3, 12, 13],
  "NL East": [15, 28, 21, 22, 20],
  "NL Central": [16, 17, 8, 23, 24],
  "NL West": [29, 27, 19, 25, 26],
};

export interface Root {
  team: Team;
}

export interface Team {
  id: string;
  uid: string;
  slug: string;
  location: string;
  name: string;
  abbreviation: string;
  displayName: string;
  shortDisplayName: string;
  color: string;
  alternateColor: string;
  isActive: boolean;
  logos: Logo[];
  record: Record;
  groups: Groups;
  links: Link[];
  franchise: Franchise;
  nextEvent: NextEvent[];
  standingSummary: string;
}

export interface Logo {
  href: string;
  width: number;
  height: number;
  alt: string;
  rel: string[];
  lastUpdated: string;
}

export interface Record {
  items: Item[];
}

export interface Item {
  description: string;
  type: string;
  summary: string;
  stats: Stat[];
}

export interface Stat {
  name: string;
  value: number;
}

export interface Groups {
  id: string;
  parent: Parent;
  isConference: boolean;
}

export interface Parent {
  id: string;
}

export interface Link {
  language: string;
  rel: string[];
  href: string;
  text: string;
  shortText: string;
  isExternal: boolean;
  isPremium: boolean;
}

export interface Franchise {
  $ref: string;
  id: string;
  uid: string;
  slug: string;
  location: string;
  name: string;
  abbreviation: string;
  displayName: string;
  shortDisplayName: string;
  color: string;
  isActive: boolean;
  venue: Venue;
  team: Team2;
  awards: Awards;
}
// ...
export interface NextEvent {
  id: string;
  date: string;
  name: string;
  shortName: string;
  season: Season;
  seasonType: SeasonType;
  week: Week;
  timeValid: boolean;
  competitions: Competition[];
  links: Link5[];
}